# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/films/schedules/schedule": {
            "post": {
                "tags": [
                    "Films",
                    "Schedules",
                    "Unfinished"
                ],
                "summary": "Create Schedule",
                "operationId": "create_schedule_films_schedules_schedule_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Schedule"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "write:schedules"
                        ]
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Films",
                    "Schedules",
                    "Unfinished"
                ],
                "summary": "Update Schedule",
                "operationId": "update_schedule_films_schedules_schedule_patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Schedule"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "write:schedules"
                        ]
                    }
                ]
            }
        },
        "/films/schedules/schedule/{schedule_id}": {
            "get": {
                "tags": [
                    "Films",
                    "Schedules",
                    "Unfinished"
                ],
                "summary": "Get Schedule",
                "operationId": "get_schedule_films_schedules_schedule__schedule_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "schedule_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Schedule Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/films/schedules/schedules": {
            "get": {
                "tags": [
                    "Films",
                    "Schedules",
                    "Unfinished"
                ],
                "summary": "Get Schedules",
                "operationId": "get_schedules_films_schedules_schedules_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Range starting ID to get"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Amount of resources to fetch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/films/film/schedules": {
            "get": {
                "tags": [
                    "Films",
                    "Unfinished"
                ],
                "summary": "Get Schedules",
                "operationId": "get_schedules_films_film_schedules_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Title"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Schedule": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "hall_id": {
                        "type": "integer",
                        "title": "Hall Id"
                    },
                    "film_id": {
                        "type": "integer",
                        "title": "Film Id"
                    },
                    "show_time": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Show Time"
                    },
                    "on_schedule": {
                        "type": "boolean",
                        "title": "On Schedule"
                    },
                    "ticket_price": {
                        "type": "number",
                        "minimum": 1.0,
                        "title": "Ticket Price"
                    },
                    "class_name": {
                        "type": "string",
                        "title": "Class Name",
                        "default": "SCHEDULE"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "hall_id",
                    "film_id",
                    "show_time",
                    "on_schedule",
                    "ticket_price"
                ],
                "title": "Schedule"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "token"
                    }
                }
            }
        }
    }
}